Replace EXE with setuptools-style entry point
The way installation of `qtsass` was done didn't work very well for me,
as it was trying to put a compiled EXE directly into `C:\Program Files`
on Windows. Instead, use the `setuptools` facility for creating
cross-platform "entry points" to install a system-wide "qtsass" command
(or at least, system-wide as long as the relevant Python directory is on
the PATH).

If a standalone EXE is required, one can always revert back to the
original `qtsass` project.